generic/cntk-lstm-forecast/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
cntk-gpu==2.7
pandas==1.4.0
yfinance==0.1.63
saxpy